#1524 refactored myTest package name
