+++ src/com/dotcms/publisher/business/PublisherQueueJob.java
@@ -13,8 +13,8 @@ import org.quartz.StatefulJob;
 import com.dotcms.publisher.business.PublishAuditStatus.Status;
 import com.dotcms.publisher.endpoint.bean.PublishingEndPoint;
 import com.dotcms.publisher.endpoint.business.PublisherEndpointAPI;
-import com.dotcms.publisher.myTest.PushPublisher;
-import com.dotcms.publisher.myTest.PushPublisherConfig;
+import com.dotcms.publisher.pusher.PushPublisher;
+import com.dotcms.publisher.pusher.PushPublisherConfig;
 import com.dotcms.publisher.util.TrustFactory;
 import com.dotmarketing.business.APILocator;
 import com.dotmarketing.exception.DotDataException;
