+++ src/com/dotcms/publisher/pusher/wrapper/PushContentWrapper.java
@@ -1,9 +1,9 @@
-package com.dotcms.publisher.myTest.wrapper;
+package com.dotcms.publisher.pusher.wrapper;
 
 import java.util.List;
 import java.util.Map;
 
-import com.dotcms.publisher.myTest.PushPublisherConfig.Operation;
+import com.dotcms.publisher.pusher.PushPublisherConfig.Operation;
 import com.dotmarketing.beans.Identifier;
 import com.dotmarketing.portlets.contentlet.model.Contentlet;
 import com.dotmarketing.portlets.contentlet.model.ContentletVersionInfo;
