+++ src/com/dotcms/publisher/receiver/handler/FolderHandler.java
@@ -7,8 +7,8 @@ import java.util.Collection;
 import java.util.Comparator;
 import java.util.List;
 
-import com.dotcms.publisher.myTest.bundler.FolderBundler;
-import com.dotcms.publisher.myTest.wrapper.FolderWrapper;
+import com.dotcms.publisher.pusher.bundler.FolderBundler;
+import com.dotcms.publisher.pusher.wrapper.FolderWrapper;
 import com.dotcms.publishing.DotPublishingException;
 import com.dotmarketing.beans.Host;
 import com.dotmarketing.beans.Identifier;
